Sorghum_Disease_Identifier_App/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pymongo==4.3.3
flask-jwt-extended==4.4.4
dnspython==2.3.0
python-dotenv==1.0.0
bcrypt==4.0.1
flask-mail==0.9.1
pyjwt==2.6.0
python-dateutil==2.8.2
email-validator==1.3.1
gunicorn==20.1.0 